requirements.txt
anyio==3.6.1
astor==0.8.1
Babel==2.10.3
bce-python-sdk==0.8.74
beautifulsoup4==4.11.1
brotlipy==0.7.0
cachetools==5.2.0
certifi
cffi
chardet==3.0.4
charset-normalizer
click==8.1.3
craft-text-detector==0.4.3
cryptography
cssselect==1.1.0
cssutils==2.6.0
cycler==0.11.0
decorator==5.1.1
einops==0.4.1
et-xmlfile==1.1.0
filelock==3.7.1
Flask==2.2.2
Flask-Babel==2.0.0
fonttools==4.34.4
future==0.18.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2022.8.1
httpcore==0.9.1
httpx==0.13.3
huggingface-hub==0.8.1
hyperframe==5.2.0
idna==2.10
imageio==2.21.0
imgaug==0.4.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.4
lmdb==1.3.0
lsd==0.3.3
lxml==4.9.1
MarkupSafe==2.1.1
matplotlib==3.5.2
mkl-fft==1.3.1
mkl-random
mkl-service==2.4.0
networkx==2.8.5
numpy
ocrd-fork-pylsd==0.0.3
opencv-contrib-python==4.4.0.46
opencv-python==4.5.4.60
openpyxl==3.0.10
opt-einsum==3.3.0
packaging==21.3
pandas==1.4.3
Pillow==9.2.0
polygon==1.0.9
Polygon3==3.0.9.1
premailer==3.10.0
protobuf==3.20.0
pyclipper==1.3.0.post3
pycocotools==2.0.4
pycparser
pycryptodome==3.15.0
pylsd==0.0.2
pyOpenSSL
pyparsing==3.0.9
PySocks
python-dateutil==2.8.2
python-Levenshtein==0.12.2
pytz==2022.1
PyWavelets==1.3.0
PyYAML==6.0
regex==2022.7.25
requests
rfc3986==1.5.0
scikit-image==0.17.2
scikit-learn==1.1.1
scipy==1.9.0
Shapely==1.8.4
six
sklearn==0.0
sniffio==1.2.0
soupsieve==2.3.2.post1
threadpoolctl==3.1.0
tifffile==2022.7.31
tokenizers==0.12.1
torch==1.12.0
torchaudio==0.12.0
torchvision==0.13.0
tqdm==4.64.0
transformers==4.21.0
typing_extensions
Unidecode==1.3.4
urllib3
visualdl==2.4.0
webcolors==1.12
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.2
zipp==3.8.1
